refactor: command 객체들을 input 패키지 안으로 이동
