+++ src/main/java/com/techfork/useraccount/application/command/UserCommandService.java
@@ -1,5 +1,8 @@
 package com.techfork.useraccount.application.command;
 
+import com.techfork.useraccount.application.command.input.CompleteOnboardingCommand;
+import com.techfork.useraccount.application.command.input.UpdateAccountProfileCommand;
+import com.techfork.useraccount.application.command.input.WithdrawUserCommand;
 import com.techfork.useraccount.domain.User;
 import com.techfork.useraccount.domain.exception.UserErrorCode;
 import com.techfork.useraccount.infrastructure.UserRepository;
