+++ src/main/java/com/techfork/useraccount/application/command/input/UpdateUserInterestsCommand.java
@@ -1,4 +1,4 @@
-package com.techfork.useraccount.application.command;
+package com.techfork.useraccount.application.command.input;
 
 import java.util.List;
